# swagger/swagger-ui/cookiehour-api.json
{
 "swagger": "2.0",
 "info": {
  "description": "API endpoints for CookieHour's implementation of the Office Hour server",
  "version": "1.0.0",
  "title": "CookieHour API",
  "termsOfService": "http://swagger.io/terms/",
  "contact": {
   "email": "apiteam@swagger.io"
  },
  "license": {
   "name": "Apache 2.0",
   "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
  }
 },
 "host": "cookiehour.swagger.io",
 "basePath": "/v2",
...
 "schemes": [
  "https",
  "http"
 ],
 "paths": {
...
  "/meeting/create/student": {
   "post": {
    "tags": [
     "meeting"
    ],
    "summary": "Creates a meeting for the logged in student",
    "description": "Schedules a meeting time slot of the chosen course's office hours",
    "operationId": "createMeeting",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for meeting create",
      "required": true,
      "schema": {
       "$ref": "#/definitions/MeetingCreate"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Operation",
      "schema": {
       "$ref": "#/definitions/MeetingData"
      },
      "headers": {
       "X-Rate-Limit": {
        "type": "integer",
        "format": "int32",
        "description": "calls per hour allowed by the user"
       },
       "X-Expires-After": {
        "type": "string",
        "format": "date-time",
        "description": "date in UTC when token expires"
       }
      }
     },
     "400": {
      "description": "Bad Request"
     },
     "404": {
      "description": "No such user - student"
     }
    }
   }
  },
...
  "/meeting/edit-agenda/student": {
   "patch": {
    "tags": [
     "meeting"
    ],
    "summary": "Edits the agenda of the student for the given meeting",
    "operationId": "editAgenda",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for agenda edit",
      "required": true,
      "schema": {
       "$ref": "#/definitions/AgendaEdit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Operation",
      "schema": {
       "$ref": "#/definitions/MeetingData"
      },
      "headers": {
       "X-Rate-Limit": {
        "type": "integer",
        "format": "int32",
        "description": "calls per hour allowed by the user"
       },
       "X-Expires-After": {
        "type": "string",
        "format": "date-time",
        "description": "date in UTC when token expires"
       }
      }
     },
     "400": {
      "description": "Bad Request"
     },
     "404": {
      "description": "No such meeting"
     }
    }
   }
  },
...
  "/meeting/edit-note/teacher": {
   "patch": {
    "tags": [
     "meeting"
    ],
    "summary": "Edits notes made in a meeting",
    "operationId": "editNote",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for note edit",
      "required": true,
      "schema": {
       "$ref": "#/definitions/NoteEdit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Operation",
      "schema": {
       "$ref": "#/definitions/MeetingData"
      },
      "headers": {
       "X-Rate-Limit": {
        "type": "integer",
        "format": "int32",
        "description": "calls per hour allowed by the user"
       },
       "X-Expires-After": {
        "type": "string",
        "format": "date-time",
        "description": "date in UTC when token expires"
       }
      }
     },
     "400": {
      "description": "Bad Request"
     },
     "404": {
      "description": "No such meeting"
     }
    }
   }
  },
  "/meeting/reschedule-time": {
   "patch": {
    "tags": [
     "meeting"
    ],
    "summary": "Reschedules the time of a meeting",
    "operationId": "editTime",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for rescheduling meeting",
      "required": true,
      "schema": {
       "$ref": "#/definitions/TimeEdit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Operation",
      "schema": {
       "$ref": "#/definitions/MeetingData"
      },
      "headers": {
       "X-Rate-Limit": {
        "type": "integer",
        "format": "int32",
        "description": "calls per hour allowed by the user"
       },
       "X-Expires-After": {
        "type": "string",
        "format": "date-time",
        "description": "date in UTC when token expires"
       }
      }
     },
     "400": {
      "description": "Bad Request"
     },
     "404": {
      "description": "No such meeting"
     }
    }
   }
  },
  "/meeting/reschedule-date": {
   "patch": {
    "tags": [
     "meeting"
    ],
    "summary": "Reschedules the date of a meeting",
    "operationId": "editDate",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for rescheduling meeting",
      "required": true,
      "schema": {
       "$ref": "#/definitions/DateEdit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Operation",
      "schema": {
       "$ref": "#/definitions/MeetingData"
      },
      "headers": {
       "X-Rate-Limit": {
        "type": "integer",
        "format": "int32",
        "description": "calls per hour allowed by the user"
       },
       "X-Expires-After": {
        "type": "string",
        "format": "date-time",
        "description": "date in UTC when token expires"
       }
      }
     },
     "400": {
      "description": "Bad Request"
     },
     "404": {
      "description": "No such meeting"
     }
    }
   }
  },
  "/meeting/delete": {
   "delete": {
    "tags": [
     "meeting"
    ],
    "summary": "Unschedules the meeting and frees up the time slot in the office hours",
    "operationId": "deleteMeeting",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "query",
      "description": "Parameters for unscheduling the meeting",
      "required": true,
      "schema": {
       "$ref": "#/definitions/DateEdit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successfully unscheduled the meeting"
     },
     "400": {
      "description": "Bad Request - invalid params"
     },
     "404": {
      "description": "Meeting not found"
     }
    }
   }
  },
  "/meeting/delete-note/teacher": {
   "delete": {
    "tags": [
     "meeting"
    ],
    "summary": "Deletes a note in a meeting",
    "operationId": "deleteNote",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for deleting a note in the meeting",
      "required": true,
      "schema": {
       "$ref": "#/definitions/DateEdit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successfully deleted the note"
     },
     "400": {
      "description": "Bad Request - invalid params"
     },
     "404": {
      "description": "Meeting not found"
     }
    }
   }
  },
  "/meeting/delete-comment": {
   "delete": {
    "tags": [
     "meeting"
    ],
    "summary": "Deletes the comment associated with the meeting",
    "operationId": "deleteComment",
    "produces": [
     "application/json"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "body",
      "description": "Parameters for deleting a comment in the meeting",
      "required": true,
      "schema": {
       "$ref": "#/definitions/DeleteComment"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successfully deleted the comment"
     },
     "400": {
      "description": "Bad Request - invalid params"
     },
     "404": {
      "description": "Meeting not found"
     }
    }
   }
  }
 },
 "securityDefinitions": {
  "cookiehour_auth": {
   "type": "oauth2",
   "authorizationUrl": "",
   "flow": "implicit",
   "scopes": {
    "write:users": "modify user information",
    "read:users": "read user information"
   }
  },
  "api_key": {
   "type": "apiKey",
   "name": "api_key",
   "in": "header"
  }
 },
 "definitions": {
...
  "MeetingCreate": {
   "type": "object",
   "properties": {
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    },
    "studentEmail": {
     "type": "string",
     "example": "student1@email.com"
    },
    "heldBy": {
     "type": "string",
     "example": "matt@email.com"
    },
    "agenda": {
     "type": "string",
     "example": "confused abt A2"
    },
    "courseName": {
     "type": "string",
     "example": "CSC302"
    }
   }
  },
  "MeetingData": {
   "type": "object",
   "properties": {
    "id": {
     "type": "integer",
     "format": "int32",
     "example": "124943593242"
    },
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    },
    "studentEmail": {
     "type": "string",
     "example": "student1@email.com"
    },
    "heldBy": {
     "type": "string",
     "example": "matt@email.com"
    },
    "agenda": {
     "type": "string",
     "example": "confused abt A2"
    },
    "courseName": {
     "type": "string",
     "example": "CSC302"
    },
    "comment": {
     "type": "string",
     "example": "some comment here"
    },
    "note": {
     "type": "string",
     "example": "secret note for me"
    }
   }
  },
  "AgendaEdit": {
   "type": "object",
   "properties": {
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    },
    "agenda": {
     "type": "string",
     "example": "confused abt A3"
    }
   }
  },
...
  "NoteEdit": {
   "type": "object",
   "properties": {
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    },
    "note": {
     "type": "string",
     "example": "secret note for myself"
    }
   }
  },
  "DeleteComment": {
   "type": "object",
   "properties": {
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    },
    "studentEmail": {
     "type": "string",
     "example": "student@email.com"
    },
    "teacherEmail": {
     "type": "string",
     "example": "matt@email.com"
    }
   }
  },
  "TimeEdit": {
   "type": "object",
   "properties": {
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    }
   }
  },
  "DateEdit": {
   "type": "object",
   "properties": {
    "date": {
     "type": "string",
     "example": "22/03/2019"
    },
    "startTime": {
     "type": "string",
     "example": "13:00"
    }
   }
  }
 },
...
}